+++ autopilot.go
@@ -9,12 +9,15 @@ import (
 	"log"
 	"net/url"
 	"os"
+	"regexp"
+	"strconv"
 	"strings"
 	"time"
 
 	"code.cloudfoundry.org/cli/cf/api/logs"
 	"code.cloudfoundry.org/cli/plugin"
 	"github.com/cloudfoundry/noaa/consumer"
+	"github.com/happytobi/autopilot/manifest"
 	"github.com/happytobi/autopilot/rewind"
 )
 
@@ -35,7 +38,7 @@ func venerableAppName(appName string) string {
 	return fmt.Sprintf("%s-venerable", appName)
 }
 
-func getActionsForApp(appRepo *ApplicationRepo, appName, manifestPath, appPath, stackName string, vars []string, varsFiles []string, showLogs bool) []rewind.Action {
+func getActionsForApp(appRepo *ApplicationRepo, appName, manifestPath, appPath, stackName string, timeout int, vars []string, varsFiles []string, envs []string, showLogs bool) []rewind.Action {
 	venName := venerableAppName(appName)
 	var err error
 	var curApp, venApp *AppEntity
@@ -97,7 +100,7 @@ func getActionsForApp(appRepo *ApplicationRepo, appName, manifestPath, appPath, 
 		// push
 		{
 			Forward: func() error {
-				return appRepo.PushApplication(appName, manifestPath, appPath, stackName, vars, varsFiles, showLogs)
+				return appRepo.PushApplication(appName, manifestPath, appPath, stackName, timeout, vars, varsFiles, envs, showLogs)
 			},
 			ReversePrevious: func() error {
 				if !haveVenToCleanup {
@@ -123,17 +126,6 @@ func getActionsForApp(appRepo *ApplicationRepo, appName, manifestPath, appPath, 
 	}
 }
 
-func getActionsForNewApp(appRepo *ApplicationRepo, appName, manifestPath, appPath, stackName string, vars []string, varsFiles []string, showLogs bool) []rewind.Action {
-	return []rewind.Action{
-		// push
-		{
-			Forward: func() error {
-				return appRepo.PushApplication(appName, manifestPath, appPath, stackName, vars, varsFiles, showLogs)
-			},
-		},
-	}
-}
-
 func (plugin AutopilotPlugin) Run(cliConnection plugin.CliConnection, args []string) {
 	// only handle if actually invoked, else it can't be uninstalled cleanly
 	if args[0] != "zero-downtime-push" {
@@ -141,11 +133,11 @@ func (plugin AutopilotPlugin) Run(cliConnection plugin.CliConnection, args []str
 	}
 
 	appRepo := NewApplicationRepo(cliConnection)
-	appName, manifestPath, appPath, stackName, vars, varsFiles, showLogs, err := ParseArgs(args)
+	appName, manifestPath, appPath, timeout, stackName, vars, varsFiles, envs, showLogs, err := ParseArgs(args)
 	fatalIf(err)
 
 	fatalIf((&rewind.Actions{
-		Actions:              getActionsForApp(appRepo, appName, manifestPath, appPath, stackName, vars, varsFiles, showLogs),
+		Actions:              getActionsForApp(appRepo, appName, manifestPath, appPath, stackName, timeout, vars, varsFiles, envs, showLogs),
 		RewindFailureMessage: "Oh no. Something's gone wrong. I've tried to roll back but you should check to see if everything is OK.",
 	}).Execute())
 
@@ -162,7 +154,7 @@ func (AutopilotPlugin) GetMetadata() plugin.PluginMetadata {
 		Version: plugin.VersionType{
 			Major: 0,
 			Minor: 0,
-			Build: 9,
+			Build: 10,
 		},
 		Commands: []plugin.Command{
 			{
@@ -187,39 +179,67 @@ func (s *StringSlice) Set(value string) error {
 	return nil
 }
 
-func ParseArgs(args []string) (string, string, string, string, []string, []string, bool, error) {
+//ParseArgs parse all cmd arguments
+func ParseArgs(args []string) (string, string, string, int, string, []string, []string, []string, bool, error) {
 	flags := flag.NewFlagSet("zero-downtime-push", flag.ContinueOnError)
 
+	var envs StringSlice
 	var vars StringSlice
 	var varsFiles StringSlice
 
 	manifestPath := flags.String("f", "", "path to an application manifest")
 	appPath := flags.String("p", "", "path to application files")
 	stackName := flags.String("s", "", "name of the stack to use")
+	timeout := flags.Int("t", 60, "push timout in secounds defualt 60s")
 	showLogs := flags.Bool("show-app-log", false, "tail and show application log during application start")
+	flags.Var(&envs, "env", "Variable key value pair for adding dynamic environment variables; can specity multiple times")
 	flags.Var(&vars, "var", "Variable key value pair for variable substitution, (e.g., name=app1); can specify multiple times")
 	flags.Var(&varsFiles, "vars-file", "Path to a variable substitution file for manifest; can specify multiple times")
 
-	if len(args) < 2 || strings.HasPrefix(args[1], "-") {
-		return "", "", "", "", []string{}, []string{}, false, ErrNoArgs
-	}
-	err := flags.Parse(args[2:])
-	if err != nil {
-		return "", "", "", "", []string{}, []string{}, false, err
+	//default start index of parameters is 2 because 1 is the appName
+	argumentStartIndex := 2
+	//if first argument is not the appName we have to read the appName out of the manifest
+	noAppNameProvided, _ := regexp.MatchString("^-[a-z]{0,3}", args[1])
+	//noAppNameProvided := strings.Contains(args[1], "-")
+	if noAppNameProvided {
+		argumentStartIndex = 1
 	}
 
-	appName := args[1]
+	err := flags.Parse(args[argumentStartIndex:])
+	if err != nil {
+		return "", "", "", *timeout, "", []string{}, []string{}, []string{}, false, err
+	}
 
 	if *manifestPath == "" {
-		return "", "", "", "", []string{}, []string{}, false, ErrNoManifest
+		return "", "", "", *timeout, "", []string{}, []string{}, []string{}, false, ErrNoManifest
 	}
 
-	return appName, *manifestPath, *appPath, *stackName, vars, varsFiles, *showLogs, nil
+	//parse first argument as appName
+	appName := args[1]
+	if noAppNameProvided {
+		manifest, err := manifest.ParseManifest(*manifestPath)
+		if err != nil {
+			return "", "", "", *timeout, "", []string{}, []string{}, []string{}, false, fmt.Errorf("error while parsing manifest %v", err)
+		}
+		appName = manifest.ApplicationManifests[0].Name
+	}
+
+	//validate var format
+	if len(envs) > 0 {
+		for _, envPair := range envs {
+			if strings.Contains(envPair, "=") == false {
+				return "", "", "", *timeout, "", []string{}, []string{}, []string{}, false, ErrWrongVarFormat
+			}
+		}
+	}
+
+	return appName, *manifestPath, *appPath, *timeout, *stackName, vars, varsFiles, envs, *showLogs, nil
 }
 
 var (
-	ErrNoArgs     = errors.New("app name must be specified")
-	ErrNoManifest = errors.New("a manifest is required to push this application")
+	ErrNoManifest     = errors.New("a manifest is required to push this application")
+	ErrManifest       = errors.New("could not parse manifest")
+	ErrWrongVarFormat = errors.New("--var would be in wrong format, use the vars like key=value")
 )
 
 type ApplicationRepo struct {
@@ -237,7 +257,7 @@ func (repo *ApplicationRepo) RenameApplication(oldName, newName string) error {
 	return err
 }
 
-func (repo *ApplicationRepo) PushApplication(appName, manifestPath, appPath, stackName string, vars []string, varsFiles []string, showLogs bool) error {
+func (repo *ApplicationRepo) PushApplication(appName, manifestPath, appPath, stackName string, timeout int, vars []string, varsFiles []string, envs []string, showLogs bool) error {
 	args := []string{"push", appName, "-f", manifestPath, "--no-start"}
 
 	if appPath != "" {
@@ -247,6 +267,10 @@ func (repo *ApplicationRepo) PushApplication(appName, manifestPath, appPath, sta
 	if stackName != "" {
 		args = append(args, "-s", stackName)
 	}
+
+	/* always append timout */
+	timeoutS := strconv.Itoa(timeout)
+	args = append(args, "-t", timeoutS)
 
 	for _, varPair := range vars {
 		args = append(args, "--var", varPair)
@@ -259,6 +283,11 @@ func (repo *ApplicationRepo) PushApplication(appName, manifestPath, appPath, sta
 	_, err := repo.conn.CliCommand(args...)
 	if err != nil {
 		return err
+	}
+
+	envErr := repo.SetEnvironmentVariables(appName, envs)
+	if envErr != nil {
+		return envErr
 	}
 
 	if showLogs {
@@ -303,6 +332,23 @@ func (repo *ApplicationRepo) PushApplication(appName, manifestPath, appPath, sta
 		return err
 	}
 
+	return nil
+}
+
+//SetEnvironmentVariable set passed envs with set-env to set variables dynamically
+func (repo *ApplicationRepo) SetEnvironmentVariables(appName string, envs []string) error {
+	varArgs := []string{"set-env", appName}
+	//set all variables passed by --var
+	for _, envPair := range envs {
+		tmpArgs := make([]string, len(varArgs))
+		copy(tmpArgs, varArgs)
+		newArgs := strings.Split(envPair, "=")
+		tmpArgs = append(tmpArgs, newArgs...)
+		_, err := repo.conn.CliCommand(tmpArgs...)
+		if err != nil {
+			return err
+		}
+	}
 	return nil
 }