+++ CHANGELOG.md
@@ -8,6 +8,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - will look for new features bringing in new versions
 - try to implement pull requests from original repository
+- make --var-files working again
+- add pushing multiapp manifests
+- add option to stop / delete old application
+- add github static page with documentation
+
+## [0.0.10] - UNRELEASED
+### Added
+- Add new argument -t to specify a push timeout
+- Add new manifest parser to get some informations out of the manifest instead of passing them through the cmd
+- Add new feature to push application without appName if manifest was provided
+- Add new argument "-env" to set-env dynamic
+- Fixtures for test
+
+### Changed
+- Changed version to 0.0.10
+- Changed ParseArgs method
+
+### Removed
+- Unused code
 
 ## [0.0.9] - 2019-01-17
 ### Added